refactor(iracing-sdk): extract telemetry snapshot formatting into shared module
Move the JSON-envelope, Markdown-report, and timestamped-filename helpers out of the telemetry-snapshot CLI into a new dependency-free snapshot.ts module (buildSnapshotEnvelope, generateMarkdown, snapshotBaseName, and the table/driver formatters), exported from the package barrel. The CLI now imports them, so the same output can be produced from inside the plugin process without a second SDK connection. Behaviour-preserving for the CLI; adds unit tests for the extracted pure functions.
